Atualizando as rotas, criando types, verificando usuario e suas requisições, erro de duplicidade, testando time do jwt, registrando cors de maneira a receber outros methods, allowedHeaders e credentials, usuarioId = user.sub (jwt)

+++ src/produtosConfig/produto.services.ts
@@ -1,29 +1,77 @@
 import { PrismaClient } from "@prisma/client";
-import { ProdutoData, ProdutoUpdateData } from "../interfaces/interfaces";
+import { ProdutoData, ProdutoDataUpdate } from "../types/types";
+import { ConflictError, NotFoundError } from "../utils/errors";
 
 const prisma = new PrismaClient();
 
 export const ProdutoService = {
   async criarProduto(data: ProdutoData) {
+    const usuario = await prisma.usuario.findUnique({
+      where: { id: data.usuarioId },
+    });
+    if (!usuario) {
+      throw new NotFoundError("Usuário não encontrado.");
+    }
     return await prisma.produto.create({ data });
   },
 
-  async editarProduto(
-    id: string,
-    data: ProdutoUpdateData
-  ) {
-    return await prisma.produto.update({ where: { id }, data });
+  async editarProduto(data: ProdutoDataUpdate) {
+    const usuario = await prisma.usuario.findUnique({
+      where: { id: data.usuarioId },
+    });
+    if (!usuario) {
+      throw new NotFoundError("Usuário não encontrado.");
+    }
+    const produto = await prisma.produto.findUnique({
+      where: {
+        id: data.id,
+      },
+    });
+    if (!produto) {
+      throw new NotFoundError("O produto não foi encontrado.");
+    }
+    return await prisma.produto.update({ where: { id: data.id }, data });
   },
 
   async listarProdutos(usuarioId: string) {
-    return await prisma.produto.findMany({ where: { usuarioId }});
+    const usuario = await prisma.usuario.findUnique({
+      where: { id: usuarioId },
+    });
+    if (!usuario) {
+      throw new NotFoundError("Usuário não encontrado.");
+    }
+    return await prisma.produto.findMany({ where: { usuarioId } });
   },
 
-  async excluirProduto(id: string) {
+  async excluirProduto(id: string, usuarioId: string) {
+    const usuario = await prisma.usuario.findUnique({
+      where: { id: usuarioId },
+    });
+    if (!usuario) {
+      throw new NotFoundError("Usuário não encontrado.");
+    }
+    const produto = await prisma.produto.findUnique({
+      where: { id },
+    });
+    if (!produto) {
+      throw new ConflictError("Produto não encontrado.");
+    }
     return await prisma.produto.delete({ where: { id } });
   },
 
-  async buscarPorId(id: string) {
+  async buscarPorId(id: string, usuarioId: string) {
+    const usuario = await prisma.usuario.findUnique({
+      where: { id: usuarioId },
+    });
+    if (!usuario) {
+      throw new NotFoundError("Usuário não encontrado.");
+    }
+    const produto = await prisma.produto.findUnique({
+      where: { id },
+    });
+    if (!produto) {
+      throw new ConflictError("Produto não encontrado.");
+    }
     return await prisma.produto.findUnique({ where: { id } });
   },
 };
